move package promutil out of sqlutil
